languages/Python/fastapi-celery/requirements.txt
celery
uvicorn
fastapi
slowapi
pydantic
redis
python-multipart
pyyaml
